Added flower animations for maps(plus, added two maps that were missing along with their camera transitions)
The map animations are all in a new folder in Backgrounds>Map Anim. The new maps were home(1C) and pathway2(2C) I wasn't sure what your code was so, of course, feel free to change the names. for the camera transitions I just duplicated one of the ones that were there and placed and renamed them accordingly.